[various] mostly tradehistory improvements

+++ xchange-hitbtc/src/main/java/org/knowm/xchange/hitbtc/v2/HitbtcAuthenticated.java
@@ -122,7 +122,7 @@ public interface HitbtcAuthenticated extends Hitbtc {
       @QueryParam("by") String sortBy,
       @QueryParam("from") String from,
       @QueryParam("till") String till,
-      @QueryParam("limit") long limit,
+      @QueryParam("limit") Integer limit,
       @QueryParam("offset") long offset)
       throws IOException, HitbtcException;
